back-end/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
black==24.10.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-jwt-auth==0.5.0
flake8==7.1.1
Flask==3.1.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.6
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pyflakes==3.2.0
PyJWT==1.7.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
Werkzeug==3.1.3
alembic==1.12.0
requests==2.25.1
numpy==1.21.0
